+++ latency/mesh.go
@@ -1,16 +1,14 @@
 package latency
 
 import (
+	"log"
 	"math/rand"
 	"net"
 	"strconv"
 	"sync"
-	"time"
 
-	"github.com/golang/protobuf/ptypes/any"
-	"github.com/hashicorp/go-hclog"
+	"github.com/maticnetwork/libp2p-gossip-bench/agent"
 	"github.com/maticnetwork/libp2p-gossip-bench/network"
-	"github.com/maticnetwork/libp2p-gossip-bench/proto"
 	ma "github.com/multiformats/go-multiaddr"
 )
 
@@ -18,33 +16,17 @@ type Mesh struct {
 	lock sync.Mutex
 
 	Latency  *LatencyData
-	Logger   hclog.Logger
-	Servers  map[string]*server
+	Logger   *log.Logger
+	Agents   map[string]*agent.Agent
 	Port     int
 	Manager  *network.Manager
 	MaxPeers int
 }
 
-func (m *Mesh) waitForPeers(min int64) {
-	check := func() bool {
-		for _, p := range m.Servers {
-			if p.NumPeers() < min {
-				return false
-			}
-		}
-		return true
-	}
-	for {
-		if check() {
-			return
-		}
-	}
-}
-
-func (m *Mesh) findPeerByPort(port string) *server {
-	for _, p := range m.Servers {
-		if strconv.Itoa(p.config.Addr.Port) == port {
-			return p
+func (m *Mesh) findPeerByPort(port string) *agent.Agent {
+	for _, a := range m.Agents {
+		if strconv.Itoa(a.Config.Addr.Port) == port {
+			return a
 		}
 	}
 	panic("bad")
@@ -66,7 +48,7 @@ func (m *Mesh) QueryLatencies(laddr, raddr ma.Multiaddr) network.Network {
 	laddrSrv := m.findPeerByPort(laddrPort)
 	raddrSrv := m.findPeerByPort(raddrPort)
 
-	latency := m.Latency.FindLatency(laddrSrv.City, raddrSrv.City)
+	latency := m.Latency.FindLatency(laddrSrv.Config.City, raddrSrv.Config.City)
 
 	nn := network.Network{
 		Kbps:    20 * 1024, // should I change this?
@@ -80,42 +62,49 @@ func (m *Mesh) Gossip(from string, size int) {
 	buf := make([]byte, size)
 	rand.Read(buf)
 
-	m.Servers[from].topic.Publish(&proto.Txn{
-		From: from,
-		Raw: &any.Any{
-			Value: buf,
-		},
-	})
+	// TODO: Check if Agent topic Publish needs Proto
+	// m.Agents[from].Topic.Publish(context.Background(), &proto.Txn{
+	// 	From: from,
+	// 	Raw: &any.Any{
+	// 		Value: buf,
+	// 	},
+	// })
 }
 
 func (m *Mesh) Join(fromID, toID string) error {
-	from := m.Servers[fromID]
-	to := m.Servers[toID]
-	return from.Join(to.AddrInfo(), 5*time.Second)
+	// from := m.Agents[fromID]
+	// to := m.Agents[toID]
+
+	// TODO: Implement Join for Agent
+	return nil
 }
 
 func (m *Mesh) RunServer(name string) error {
-	if m.Servers == nil {
-		m.Servers = map[string]*server{}
+	if m.Agents == nil {
+		m.Agents = map[string]*agent.Agent{}
 	}
 
 	n := rand.Int() % len(m.Latency.SourcesList)
 	city := m.Latency.SourcesList[n].Name
 
-	config := network.DefaultConfig()
+	config := agent.DefaultConfig()
 	config.Transport = m.Manager.Transport()
 	config.City = city
-	config.MaxPeers = uint64(m.MaxPeers)
+	config.MaxPeers = int64(m.MaxPeers)
 
 	config.Addr = &net.TCPAddr{IP: net.ParseIP("127.0.0.1"), Port: m.Port}
 	m.Port++
 
-	m.Servers[name] = newServer(m.Logger.Named(name), config, city)
+	agent, err := agent.NewAgent(m.Logger, config)
+	if err != nil {
+		return err
+	}
+	m.Agents[name] = agent
 	return nil
 }
 
 func (m *Mesh) Stop() {
-	for _, srv := range m.Servers {
-		srv.Close()
+	for _, agnt := range m.Agents {
+		agnt.Stop()
 	}
 }
